Avoid std::call_once (GH #707) This commit also favors init priorities over C++ dynamic initialization. After the std::call_once problems on Sparc and PowerPC I'm worried about problems with Dynamic Initialization and Destruction with Concurrency. We also do away with supressing warnings and use CRYPTOPP_UNUSED instead.

+++ blake2.cpp
@@ -235,7 +235,7 @@ void BLAKE2_Base<word32, false>::UncheckedSetKey(const byte *key, unsigned int l
         AlignedSecByteBlock temp(BLOCKSIZE);
         memcpy_s(temp, BLOCKSIZE, key, length);
 
-        size_t rem = SaturatingSubtract(BLOCKSIZE, length);
+        size_t rem = SaturatingSubtract((unsigned int)BLOCKSIZE, length);
         if (rem)
             std::memset(temp+length, 0x00, rem);
 
@@ -291,7 +291,7 @@ void BLAKE2_Base<word64, true>::UncheckedSetKey(const byte *key, unsigned int le
         AlignedSecByteBlock temp(BLOCKSIZE);
         memcpy_s(temp, BLOCKSIZE, key, length);
 
-        size_t rem = SaturatingSubtract(BLOCKSIZE, length);
+        size_t rem = SaturatingSubtract((unsigned int)BLOCKSIZE, length);
         if (rem)
             std::memset(temp+length, 0x00, rem);
